requirements.txt
########## Core / CLI ##########
click==8.1.8            # ← bumped from 8.1.7 to satisfy duckduckgo-search
typer==0.12.3
rich==13.7.1
requests==2.31.0
httpx==0.27.0
aiohttp==3.9.5
tqdm==4.66.4
loguru==0.7.2
schedule==1.2.1
python-dotenv==1.0.1
watchdog==4.0.0
markdown==3.6 # NEW: Added markdown explicitly

########## Browser / Scraping ##########
playwright==1.44.0
selenium==4.21.0
beautifulsoup4==4.12.3
lxml==5.3.1    # ← bump from 5.2.1 to satisfy duckduckgo-search

########## LLM / RAG ##########
torch==2.3.0
torchvision==0.18.0
sentencepiece==0.2.0
tokenizers==0.19.1
transformers==4.41.1
accelerate==0.30.0
sentence-transformers==2.7.0
langchain==0.2.1
llama-index==0.11.8
faiss-cpu==1.11.0

########## PDF / OCR / Office ##########
PyPDF2==3.0.1
pdfminer.six==20221105
weasyprint==61.2
reportlab==4.0.8
ocrmypdf==16.1.0
python-docx==1.1.0
openpyxl==3.1.2
pillow==10.3.0
pytesseract==0.3.10
opencv-python-headless==4.11.0.86

########## Speech / Audio ##########
openai-whisper==20231117
faster-whisper==1.1.1
webrtcvad-wheels==2.0.14
sounddevice==0.4.6
soundfile==0.12.1
pyttsx3==2.90

########## Data / Utilities ##########
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
scikit-learn==1.5.0
networkx==3.3
orjson==3.10.3
ujson==5.9.0
duckduckgo-search==8.0.5     # DuckDuckGo text/news/images/video search
dnspython==2.6.1
python-whois==0.9.5
GitPython==3.1.43


########## ASGI / API ##########
fastapi==0.111.0
uvicorn[standard]==0.30.0
